+++ HealthActivity/Flow/TabBar/ProfileFlow/Profile/ProfileViewModel.swift
@@ -59,7 +59,7 @@ struct ProfileViewModel: BaseViewModelType {
         healthManager.getHeartRate { state in
             switch state {
             case .success(let heartModel):
-                self.heartRateSubject.onNext(heartModel.value)
+                self.heartRateSubject.onNext(heartModel.value.toInt)
             case .failure(let error):
                 self.errorSubject.onNext(error)
             }
